pyproject.toml
[project]
name = "reason-sot"
version = "0.1.0"
description = "Low-latency dual-process reasoning engine for voice interview agents"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.52.0",
    "pydantic>=2.7.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: tests that hit the live Anthropic API (deselect with -m 'not integration')",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.backends._legacy:_Backend"
